+++ NestSuite/Converters/WorkspaceKindToAccentBrushConverter.cs
@@ -6,18 +6,30 @@ namespace NestSuite.Converters;
 
 public class WorkspaceKindToAccentBrushConverter : IValueConverter
 {
+    private static readonly SolidColorBrush NoteNestBrush = CreateFrozenBrush(0x4A, 0x90, 0xD9);
+    private static readonly SolidColorBrush IdeaNestBrush = CreateFrozenBrush(0xE8, 0xA0, 0x20);
+    private static readonly SolidColorBrush ChatNestBrush = CreateFrozenBrush(0x4C, 0xAF, 0x50);
+    private static readonly SolidColorBrush TempBrush = CreateFrozenBrush(0xA0, 0xA0, 0xA8);
+
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         => value is NestSuiteWorkspaceKind kind
             ? kind switch
             {
-                NestSuiteWorkspaceKind.NoteNest => new SolidColorBrush(Color.FromRgb(0x4A, 0x90, 0xD9)),
-                NestSuiteWorkspaceKind.IdeaNest => new SolidColorBrush(Color.FromRgb(0xE8, 0xA0, 0x20)),
-                NestSuiteWorkspaceKind.ChatNest  => new SolidColorBrush(Color.FromRgb(0x4C, 0xAF, 0x50)),
-                NestSuiteWorkspaceKind.Temp => new SolidColorBrush(Color.FromRgb(0xA0, 0xA0, 0xA8)),
+                NestSuiteWorkspaceKind.NoteNest => NoteNestBrush,
+                NestSuiteWorkspaceKind.IdeaNest => IdeaNestBrush,
+                NestSuiteWorkspaceKind.ChatNest  => ChatNestBrush,
+                NestSuiteWorkspaceKind.Temp => TempBrush,
                 _ => Brushes.Transparent
             }
             : (object)Brushes.Transparent;
 
     public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        => throw new NotSupportedException();
+        => Binding.DoNothing;
+
+    private static SolidColorBrush CreateFrozenBrush(byte red, byte green, byte blue)
+    {
+        var brush = new SolidColorBrush(Color.FromRgb(red, green, blue));
+        brush.Freeze();
+        return brush;
+    }
 }