+++ NestSuite/NestSuite/IdeaNest/Converters/IdeaStringIsEmptyToVisibilityConverter.cs
@@ -14,5 +14,5 @@ public class IdeaStringIsEmptyToVisibilityConverter : IValueConverter
     }
 
     public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        => throw new NotSupportedException();
+        => Binding.DoNothing;
 }